Rename value-object to types
