+++ application/repositories/StockRepository.ts
@@ -1,5 +1,5 @@
 import { Stock } from "../../domain/entities/Stock";
-import { StockQuantity } from "../../domain/value-object/stock";
+import { StockQuantity } from "../../domain/types/stock";
 
 export interface StockRepository {
   addStock(sparePartId: string, quantity: StockQuantity): Promise<Stock>;
