+++ domain/types/fuel-capacity.ts
