+++ domain/types/fuel-type.ts
