+++ domain/types/motorcycle-type.ts
