+++ domain/types/replaced-part.ts
