+++ domain/types/stock.ts
